refactor: move copy/mirror transfer constants to @/drizzle/constants per review and rename the stale "halves" drain test to match its divide-by-duration behavior
