+++ app/src/libs/combat/constants.ts
@@ -172,95 +172,3 @@ export const POST_PIERCE_TAGS: string[] = [
   "decreaseheal",
   "increaseheal",
 ];
-
-/**
- * Maximum number of effects the `copy` / `mirror` tags may transfer. This is a
- * PERSISTENT, per-caster ceiling: at most this many copied effects may be active
- * on the caster (mirrored on the target) at once, across any number of casts.
- */
-export const COPY_MAX_TAGS = 4;
-export const MIRROR_MAX_TAGS = 4;
-
-/**
- * Commitment-priority tiers for the `copy` tag (opponent buffs -> self), highest
- * commitment first. The array index of a tier IS its rank. Same-tier types are
- * independent slots, tie-broken by effect power.
- *
- * NOTE: This is unrelated to `sortEffects` round-application ordering. These tiers
- * only govern copy/mirror *selection* — which effects are eligible and which win
- * the capped slots.
- */
-export const COPY_PRIORITY_TIERS: readonly (readonly string[])[] = [
-  ["increasedamagegiven", "decreasedamagetaken"],
-  ["lifesteal"],
-  ["absorb", "reflect"],
-  ["shield"],
-  ["increaseheal"],
-  ["increasestat"],
-];
-
-/**
- * Commitment-priority tiers for the `mirror` tag (own debuffs -> target). Only the
- * named types are ranked; any other mirror-eligible negative type falls to a tail
- * rank (supplied by the caller) and is tie-broken by power.
- */
-export const MIRROR_PRIORITY_TIERS: readonly (readonly string[])[] = [
-  ["increasedamagetaken", "decreasedamagegiven"],
-  ["afterburn"],
-  ["poison", "increasepoolcost"],
-  ["wound"],
-];
-
-/** Effect types the `copy` tag may transfer (flattened copy priority tiers). */
-export const COPYABLE_EFFECT_TYPES: string[] = COPY_PRIORITY_TIERS.flat();
-
-/**
- * Source (`fromType`) origins that copy/mirror never transfer — passive/gear
- * effects are not stealable/reflectable. Shared by both tags.
- */
-export const TRANSFER_EXCLUDED_SOURCE_TYPES: string[] = [
-  "bloodline",
-  "armor",
-  "accessory",
-  "keystone",
-  "item",
-  "village",
-  "skill",
-];
-
-/**
- * Negative effect TYPES the `mirror` tag never reflects onto the target (raw damage,
- * hard control, and non-transferable mechanics). This is the effect-type axis
- * (`e.type`) — distinct from `TRANSFER_EXCLUDED_SOURCE_TYPES`, which excludes by
- * source origin (`e.fromType`); the two lists are not interchangeable.
- * `wound` is intentionally NOT here — it is mirror-eligible.
- */
-export const MIRROR_EXCLUDED_EFFECT_TYPES: string[] = [
-  "damage",
-  "pierce",
-  "clear",
-  "buffprevent",
-  "cleanseprevent",
-  "moveprevent",
-  "healprevent",
-  "timecompression",
-];
-
-/** Build a `type -> rank` lookup (lower rank = higher priority) from tier arrays. */
-const buildTagPriorityRank = (
-  tiers: readonly (readonly string[])[],
-): ReadonlyMap<string, number> => {
-  const rank = new Map<string, number>();
-  for (let index = 0; index < tiers.length; index++) {
-    const tier = tiers[index];
-    if (tier) {
-      for (const type of tier) {
-        rank.set(type, index);
-      }
-    }
-  }
-  return rank;
-};
-
-export const COPY_PRIORITY_RANK = buildTagPriorityRank(COPY_PRIORITY_TIERS);
-export const MIRROR_PRIORITY_RANK = buildTagPriorityRank(MIRROR_PRIORITY_TIERS);
